pyproject.toml
[project]
name = "mcp-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "google-api-python-client>=2.174.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.23",
    "mcp-use>=1.3.1",
    "mcp[cli]>=1.9.4",
    "mongo-mcp>=0.1.3",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
